Delete komento omassa luokassaan
