+++ src/main/java/projekti/UI/commands/Command.java
@@ -4,4 +4,5 @@ import java.sql.SQLException;
 
 public interface Command {
     public void execute() throws SQLException;
+
 }
