+++ src/main/java/projekti/UI/commands/CreateRecommendation.java
@@ -20,7 +20,7 @@ public class CreateRecommendation implements Command {
         this.rh = rh;
         this.db = db;
     }
-
+    @Override
     public void execute() throws SQLException {
         String recommendationType = rh.askType().toLowerCase();
         Function<Property, String> requestProperty = (Property property) -> {
@@ -45,7 +45,9 @@ public class CreateRecommendation implements Command {
             rh.getIO().println("\nrecommendation not added");
         }
     }
-   
+ 
+ 
+
     private Recommendation save(Recommendation recommendation) throws SQLException {
         switch (recommendation.getType()) {
             case "BOOK":
@@ -57,5 +59,5 @@ public class CreateRecommendation implements Command {
             default:
                 throw new IllegalArgumentException("No save definition for recommendation of type: " + recommendation.getType());
         }
-    }
-}  
+    } 
+}
