+++ src/main/java/projekti/UI/commands/SelectRecommendation.java
@@ -2,11 +2,13 @@ package projekti.UI.commands;
 
 import java.sql.SQLException;
 
+import projekti.domain.Recommendation;
+
 public class SelectRecommendation implements Command {
 
     @Override
     public void execute() throws SQLException {
-		 //coming soon
-	}
-   
+        // coming soon
+    }
+
 }
