+++ src/main/java/projekti/UI/commands/UpdateRecommendation.java
@@ -2,11 +2,14 @@ package projekti.UI.commands;
 
 import java.sql.SQLException;
 
+import projekti.domain.Recommendation;
+
 public class UpdateRecommendation implements Command {
 
     @Override
     public void execute() throws SQLException {
-		 //coming soon
-	}
+        // coming soon
+    }
+
    
 }
